# nalanda-enterprise-training-platform/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nalanda Training and Development API
  version: 1.0.0
  description: API contract for role-based learning, course management, strict assessments, proctoring, analytics, and reports.
servers:
  - url: http://localhost:8080/api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [Admin, Manager, User] }
        department: { type: string }
        managerId: { type: string, nullable: true }
        status: { type: string, enum: [Active, Inactive] }
    Course:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        skill: { type: string }
        tags:
          type: array
          items: { type: string }
        approval: { type: string, enum: [Pending, Approved, Rejected] }
        status: { type: string, enum: [Active, Inactive] }
        version: { type: integer }
    Assessment:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        courseId: { type: string }
        type: { type: string, enum: [MCQ, Descriptive, Timed Quiz] }
        strictMode: { type: boolean }
        approval: { type: string, enum: [Pending, Approved, Rejected] }
paths:
  /health:
    get:
      security: []
      summary: Service health
      responses:
        "200": { description: OK }
  /auth/login:
    post:
      security: []
      summary: Login and receive JWT
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        "200": { description: Authenticated }
        "401": { description: Invalid credentials }
  /auth/me:
    get:
      summary: Current authenticated user
      responses:
        "200": { description: Current user }
  /users:
    get:
      summary: List users visible to Admin or Manager
      parameters:
        - in: query
          name: department
          schema: { type: string }
        - in: query
          name: role
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200": { description: User list }
    post:
      summary: Create user; Admin only
      responses:
        "201": { description: User created }
  /users/{id}:
    patch:
      summary: Update user role, manager, department, or profile fields
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: User updated }
  /users/{id}/status:
    patch:
      summary: Inactivate or reactivate user without deleting records
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Status changed }
  /courses:
    get:
      summary: List courses by role and filters
      responses:
        "200": { description: Course list }
    post:
      summary: Create course with PDF, text, video link, or uploaded video content
      responses:
        "201": { description: Course created }
  /courses/{id}:
    put:
      summary: Update course and create a new version
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Course updated }
  /courses/{id}/duplicate:
    post:
      summary: Duplicate a course as a new pending draft
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "201": { description: Course duplicated }
  /courses/{id}/approval:
    patch:
      summary: Admin approval or rejection for course
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Course reviewed }
  /assignments:
    post:
      summary: Assign course to individuals or teams
      responses:
        "201": { description: Assignment created }
  /assessments:
    get:
      summary: List assessments by role
      responses:
        "200": { description: Assessment list }
    post:
      summary: Create assessment and submit for approval when created by Manager
      responses:
        "201": { description: Assessment created }
  /assessments/{id}/approval:
    patch:
      summary: Admin approval or rejection for assessment
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Assessment reviewed }
  /assessments/{id}/attempts/start:
    post:
      summary: Start an assessment attempt with strict mode rules
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "201": { description: Attempt started }
  /assessments/attempts/{attemptId}/submit:
    post:
      summary: Submit answers and calculate MCQ score or queue manual review
      parameters:
        - in: path
          name: attemptId
          required: true
          schema: { type: string }
      responses:
        "200": { description: Attempt submitted }
  /proctoring/events:
    post:
      summary: Append proctoring events such as tab switch, noise, camera off, or fullscreen exit
      responses:
        "201": { description: Proctoring event logged }
  /analytics/global:
    get:
      summary: Admin global analytics
      responses:
        "200": { description: Analytics returned }
  /analytics/team:
    get:
      summary: Manager team analytics
      responses:
        "200": { description: Analytics returned }
  /analytics/me:
    get:
      summary: User learning analytics
      responses:
        "200": { description: Analytics returned }
  /reports/export:
    post:
      summary: Queue PDF or Excel report export
      responses:
        "202": { description: Report queued or ready }
  /settings:
    get:
      summary: Read proctoring and security settings
      responses:
        "200": { description: Settings returned }
    patch:
      summary: Update proctoring and security settings
      responses:
        "200": { description: Settings updated }
  /audit-logs:
    get:
      summary: Admin audit log feed
      responses:
        "200": { description: Audit logs returned }